backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
flask==3.0.0
flask-cors==4.0.0
flask-socketio==5.3.6
requests==2.31.0
mysql-connector-python==8.2.0
PyMySQL==1.1.0
pandas==2.1.4
python-dotenv==1.0.0
selenium==4.15.2
openpyxl==3.1.2
xlsxwriter==3.1.9
psutil==5.9.6
websockets==12.0
python-multipart==0.0.6
jinja2==3.1.2
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
cryptography==41.0.8
setuptools==69.0.3